Refactor resolver cache tests to use async operations and organize in resolver/ directory
- Move tests from test/bundler/resolver_cache_*.test.ts to test/bundler/resolver/cache-*.test.ts
- Refactor all tests to use async fs operations (fs.promises) instead of sync
- Replace Bun.spawnSync with Bun.spawn using await using for automatic cleanup
- Add describe.concurrent to runtime and Node compat tests for faster execution
- Remove unnecessary async keywords from test functions that don't need them
- Remove duplicate test in runtime tests
- Reduce console.log verbosity in Node compat tests (only log on failure)
- Add output verification for all cache invalidation tests

Tests now properly verify the resolver cache invalidation fix works correctly.
System bun fails 2 tests as expected (confirms the bug exists), debug build passes all 13 tests.
